+++ src/chains-config/westendControllers.ts
@@ -33,6 +33,7 @@ export const westendControllers: ControllerConfig = {
 		'NodeNetwork',
 		'NodeTransactionPool',
 		'NodeVersion',
+		'PalletsDispatchables',
 		'PalletsConsts',
 		'PalletsErrors',
 		'PalletsEvents',