+++ src/controllers/pallets/index.ts
@@ -16,6 +16,7 @@
 
 export { default as PalletsAssets } from './PalletsAssetsController';
 export { default as PalletsConsts } from './PalletsConstsController';
+export { default as PalletsDispatchables } from './PalletsDispatchablesController';
 export { default as PalletsErrors } from './PalletsErrorsController';
 export { default as PalletsEvents } from './PalletsEventsController';
 export { default as PalletsNominationPools } from './PalletsNominationPoolsController';