+++ src/main.ts
@@ -90,7 +90,10 @@ async function main() {
 	});
 
 	// Start the server
-	app.listen();
+	const server = app.listen();
+
+	server.keepAliveTimeout = config.EXPRESS.KEEP_ALIVE_TIMEOUT;
+	server.headersTimeout = config.EXPRESS.KEEP_ALIVE_TIMEOUT + 5000;
 
 	if (args.prometheus) {
 		// Create Metrics App